pyproject.toml
[project]
name = "market-analysis-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "langgraph==0.3.21",
    "langchain-openai==0.3.12",
    "pydantic-settings==2.9.1",
    "beautifulsoup4==4.13.4",
    "reportlab==4.4.1",
    "matplotlib==3.10.1",
    "aiofiles==24.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "ruff>=0.9.0",
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = ["app/prompts.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
known-third-party = ["tests"]
